LayoutTests/compositing/updates/no-style-change-updates-expected.txt

Test that compositing updates do not happen for style changes that do not need them.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS updateCount is 0
PASS updateCount is 1
PASS updateCount is 0
PASS updateCount is 1
PASS updateCount is 0
PASS successfullyParsed is true

TEST COMPLETE
